refactor: refactor format sheet data
